// src/shared/types/contact.ts
export interface MultiValueContactField {
  value: string;
  type: string;
  label: string;
}

export interface SignificantDateField {
  value: string;
  type: string;
  label: string;
}

export interface RelatedPersonField {
  value: string;
  type: string;
  label: string;
}

export interface CustomContactField {
  key: string;
  value: string;
}